+++ src/components/shared/check/index.ts
@@ -1,0 +1,1 @@
+export { Check } from './check';
