--- src/components/shared/three-state-checkbox/index.ts
@@ -1,1 +1,0 @@
-export { ThreeStateCheckbox } from './three-state-checkbox';
